+++ src/handler/endpoint.rs
@@ -1,30 +1,42 @@
-use crate::{description, di::Injectable, from_fn_with_description, Handler, HandlerDescription};
+use crate::{
+    description, di::Injectable, from_fn_with_description, Handler, HandlerDescription,
+    HandlerSignature,
+};
 use futures::FutureExt;
-use std::{ops::ControlFlow, sync::Arc};
+use std::{collections::HashSet, ops::ControlFlow, sync::Arc};
 
 /// Constructs a handler that has no further handlers in a chain.
 ///
 /// An endpoint is a handler that _always_ breaks handler execution after its
 /// completion. So, you can use it when your chain of responsibility must end
 /// up, and handle an incoming event.
+///
+/// # Signature
+///
+/// The run-time type signature of this handler is `HandlerSignature::Other {
+/// input_types: F::input_types(), output_types: HashSet::from([]) }`.
 #[must_use]
 #[track_caller]
 pub fn endpoint<'a, F, Input, Output, FnArgs, Descr>(f: F) -> Endpoint<'a, Input, Output, Descr>
 where
     F: Injectable<Input, Output, FnArgs> + Send + Sync + 'a,
     Input: Send + 'a,
-    Output: 'a,
+    Output: 'static,
     Descr: HandlerDescription,
 {
     let f = Arc::new(f);
 
-    from_fn_with_description(Descr::endpoint(), move |x, _cont| {
-        let f = Arc::clone(&f);
-        async move {
-            let f = f.inject(&x);
-            f().map(ControlFlow::Break).await
-        }
-    })
+    from_fn_with_description(
+        Descr::endpoint(),
+        move |x, _cont| {
+            let f = Arc::clone(&f);
+            async move {
+                let f = f.inject(&x);
+                f().map(ControlFlow::Break).await
+            }
+        },
+        HandlerSignature::Other { input_types: F::input_types(), output_types: HashSet::from([]) },
+    )
 }
 
 /// A handler with no further handlers in a chain.